[ICU-6799] fix compiler warnings X-Commit-URL: https://ssl.icu-project.org/trac/changeset/25668 Former-commit-id: 2ee1bd56cf05d70caacdf2783e3fcc0ca4e2b6c5

+++ icu4c/source/common/usprep.cpp
@@ -443,7 +443,7 @@ usprep_openByType(UStringPrepProfileType type,
         return NULL;
     }
     int32_t index = (int32_t)type;
-    if (index < 0 || index >= sizeof(PROFILE_NAMES)/sizeof(PROFILE_NAMES[0])) {
+    if (index < 0 || index >= (int32_t)(sizeof(PROFILE_NAMES)/sizeof(PROFILE_NAMES[0]))) {
         *status = U_ILLEGAL_ARGUMENT_ERROR;
         return NULL;
     }
